requirements.txt
# Libraries required for Gramex

# General utilities
pathlib                     # Python 3.3+ already has it
orderedattrdict             # To load configs
zope.dottedname             # To resolve Python names into objects

# Application specific
tornado                     # For the app
PyYAML                      # For config loading
crontab                     # For scheduler
